+++ CHANGES.rst
@@ -8,6 +8,7 @@ Improvements:
  - Improve certificate pinning management for HomeServerConnectionConfig.
 
 Bugfix:
+ - Fix strip previous reply when they contain new line (vector-im/riot-android#2612)
  - Enable CLEARTEXT communication for http endpoints (vector-im/riot-android#2495)
 
 API Change: